Create Program 2
Write a program using if-blocks that allows the user to input a taxable income value into a text box, and after clicking a button the tax is calculated and displayed in another text box.  Write a function to do the tax calculation.  The answer should be rounded to 2 decimal places.  Your program should calculate the correct tax no matter what the income range.  Use an IF-block to make this work.