Create Program 3
In this program, when a button is clicked a function is called that generates an html table of miles per hour values and their
equivalent kilometers per hour values on the page. The table should be inserted in an HTML divison. Use the array of mph values given below. You cannot modify this array. The kph values
should be rounded to 1 decimal place.