Create Program 4
Your program will take this data and generate an HTML table to display it to the user. When you first run your code it should display something similar the following heading and button. When you press press the button to display your code should display a Bootstrap styled table of the beach items along with the calculated total for each object. For example, the first item, boogie board, 3 were purchased. So the subtotal is 3 * 15.99 = 47.97. The first 2 rows of the table should look like this. The table should have 6 rows in all.