Create Program 5
You will need to use a Handlebars template that includes an #each section to format the displayed data into an HTML table.

You will need to upload the index.html file along with the cit5students.json (unmodified) file to your github website so that the program can be run from there.

Your index.html page should make an Ajax request that requests the JSON data and then converts the JSON data to a regular JavaScript array of objects. This array of objects should then populate the Handlebars template.